Wed, Aug 23, 2023, 4:01 AM +03:00

+++ finance/app.py
@@ -80,7 +80,7 @@ def buy():
 
         result = db.execute("SELECT cash FROM users WHERE id = ?", session["user_id"])
         if result:
-            cash = int(result[0]["cash"])
+            cash = float(result[0]["cash"])
         else:
             return apology("oops", 400)
